Backporting change from main Antlr4 repo
original commit message:

[Swift] Change getStateToAltMap to use Int as the dictionary key, not ATNState.

The ATNState hashValue and == override are just using the stateNumber field, so
using the Int directly is equivalent, and saves bouncing through those methods.

This also seems to be a correctness issue with the new Hashable protocol changes in
Swift 4.2 (though I admit that I don't know why).

Remove PredictionMode.getStateToAltMap, which was just a stub onto
ATNConfigSet.getStateToAltMap and didn't seem to be doing anything useful.

ref: https://github.com/antlr/antlr4/commit/8cd798431d80fd5f2363a715e04b3c57cfcf4815#diff-3698841a8103c0d6b28854889b5c2f40